Split preset files
